Reuse const rendering from rustdoc in rmeta encoding
